frontend adjustments
